requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
blas=1.0=mkl
brotli-python=1.0.9=py310hd77b12b_7
bzip2=1.0.8=h2bbff1b_5
ca-certificates=2024.3.11=haa95532_0
certifi=2024.2.2=py310haa95532_0
cffi=1.16.0=py310h2bbff1b_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
colorama=0.4.6=pypi_0
contourpy=1.2.1=pypi_0
cuda-cccl=12.4.127=0
cuda-cudart=12.1.105=0
cuda-cudart-dev=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-libraries-dev=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvrtc-dev=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.4.127=0
cuda-opencl-dev=12.4.127=0
cuda-profiler-api=12.4.127=0
cuda-runtime=12.1.0=0
cudatoolkit=11.8.0=hd77b12b_0
cycler=0.12.1=pypi_0
eigen=3.4.0=h59b6b97_0
ffmpeg=4.2.2=he774522_0
filelock=3.13.1=py310haa95532_0
fonttools=4.51.0=pypi_0
freetype=2.12.1=ha860e81_0
fsspec=2023.10.0=py310haa95532_0
giflib=5.2.1=h8cc25b3_3
glib=2.78.4=hd77b12b_0
glib-tools=2.78.4=hd77b12b_0
gmpy2=2.1.2=py310h7f96b67_0
gst-plugins-base=1.18.5=h9e645db_0
gstreamer=1.18.5=hd78058f_0
hdf5=1.12.1=h51c971a_3
icc_rt=2022.1.0=h6049295_2
icu=58.2=ha925a31_3
idna=3.4=py310haa95532_0
intel-openmp=2023.1.0=h59b6b97_46320
jinja2=3.1.3=py310haa95532_0
jpeg=9e=h2bbff1b_1
kiwisolver=1.4.5=pypi_0
krb5=1.20.1=h5b6d351_0
lerc=3.0=hd77b12b_0
libclang=14.0.6=default_hb5a9fac_1
libclang13=14.0.6=default_h8e68704_1
libcublas=12.1.0.26=0
libcublas-dev=12.1.0.26=0
libcufft=11.0.2.4=0
libcufft-dev=11.0.2.4=0
libcurand=10.3.5.147=0
libcurand-dev=10.3.5.147=0
libcusolver=11.4.4.55=0
libcusolver-dev=11.4.4.55=0
libcusparse=12.0.2.55=0
libcusparse-dev=12.0.2.55=0
libdeflate=1.17=h2bbff1b_1
libffi=3.4.4=hd77b12b_0
libglib=2.78.4=ha17d25a_0
libiconv=1.16=h2bbff1b_2
libjpeg-turbo=2.0.0=h196d8e1_0
libnpp=12.0.2.50=0
libnpp-dev=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjitlink-dev=12.1.105=0
libnvjpeg=12.1.1.14=0
libnvjpeg-dev=12.1.1.14=0
libogg=1.3.5=h2bbff1b_1
libpng=1.6.39=h8cc25b3_0
libpq=12.17=h906ac69_0
libprotobuf=3.20.3=h23ce68f_0
libtiff=4.5.1=hd77b12b_0
libuv=1.44.2=h2bbff1b_0
libvorbis=1.3.7=he774522_0
libwebp=1.3.2=hbc33d0d_0
libwebp-base=1.3.2=h2bbff1b_0
libxml2=2.10.4=h0ad7f3c_1
libxslt=1.1.37=h2bbff1b_1
lz4-c=1.9.4=h2bbff1b_0
markupsafe=2.1.3=py310h2bbff1b_0
matplotlib=3.8.4=pypi_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py310h2bbff1b_1
mkl_fft=1.3.8=py310h2bbff1b_0
mkl_random=1.2.4=py310h59b6b97_0
mpc=1.1.0=h7edee0f_1
mpfr=4.0.2=h62dcd97_1
mpir=3.0.0=hec2e145_1
mpmath=1.3.0=py310haa95532_0
networkx=3.1=py310haa95532_0
ninja=1.10.2=haa95532_5
ninja-base=1.10.2=h6d14046_5
numpy=1.26.4=py310h055cbcc_0
numpy-base=1.26.4=py310h65a83cf_0
opencv=4.6.0=py310ha36de5b_5
opencv-python=4.9.0.80=pypi_0
openjpeg=2.4.0=h4fc8c34_0
openssl=3.0.13=h2bbff1b_0
packaging=24.0=pypi_0
pandas=2.2.2=pypi_0
pcre2=10.42=h0ff8eda_0
pillow=10.2.0=py310h2bbff1b_0
pip=23.3.1=py310haa95532_0
psutil=5.9.8=pypi_0
py-cpuinfo=9.0.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pyparsing=3.1.2=pypi_0
pysocks=1.7.1=py310haa95532_0
python=3.10.14=he1021f5_0
python-dateutil=2.9.0.post0=pypi_0
pytorch=2.2.1=py3.10_cuda12.1_cudnn8_0
pytorch-cuda=12.1=hde6ce7c_5
pytorch-mutex=1.0=cuda
pytz=2024.1=pypi_0
pyyaml=6.0.1=py310h2bbff1b_0
qt-main=5.15.2=h879a1e9_9
qt-webengine=5.15.9=h5bd16bc_7
qtwebkit=5.212=h2bbfb41_5
requests=2.31.0=py310haa95532_1
scipy=1.13.0=pypi_0
seaborn=0.13.2=pypi_0
setuptools=68.2.2=py310haa95532_0
six=1.16.0=pypi_0
sqlite=3.41.2=h2bbff1b_0
sympy=1.12=py310haa95532_0
tbb=2021.8.0=h59b6b97_0
thop=0.1.1-2209072238=pypi_0
tk=8.6.12=h2bbff1b_0
torchaudio=2.2.1=pypi_0
torchvision=0.17.1=pypi_0
tqdm=4.66.2=pypi_0
typing_extensions=4.9.0=py310haa95532_1
tzdata=2024.1=pypi_0
ultralytics=8.2.0=pypi_0
urllib3=2.1.0=py310haa95532_1
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.41.2=py310haa95532_0
win_inet_pton=1.1.0=py310haa95532_0
xz=5.4.6=h8cc25b3_0
yaml=0.2.5=he774522_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.5=hd43e919_0
